add products to public, restricted seed, CSF - home

+++ pages/api/seed.ts
@@ -11,11 +11,11 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<Data>
 ) {
-  // if (process.env.NODE_ENV === 'production') {
-  //   return res
-  //     .status(401)
-  //     .json({ message: 'You do NOT have an access to this route' });
-  // }
+  if (process.env.NODE_ENV === 'production') {
+    return res
+      .status(401)
+      .json({ message: 'You do NOT have an access to this route' });
+  }
 
   await db.connect();
   await User.deleteMany();
